Create class RecordYear and update class FileUtility

+++ src/ru/pls/Record.java
@@ -1,14 +1,12 @@
 package ru.pls;
 
-import java.math.BigDecimal;
-
 public class Record {
-    String itemName;
-    Boolean isExpense;
-    Integer quantity;
-    BigDecimal sumOfOne;
+   String itemName;
+   Boolean isExpense;
+   Integer quantity;
+   Integer sumOfOne;
 
-    public Record(String itemName, Boolean isExpense, Integer quantity, BigDecimal sumOfOne) {
+    public Record(String itemName, Boolean isExpense, Integer quantity, Integer sumOfOne) {
         this.itemName = itemName;
         this.isExpense = isExpense;
         this.quantity = quantity;
